requirements.txt
beautifulsoup4==4.11.1
PyYAML==6.0
aiohttp==3.8.1
pyTelegramBotAPI==4.7.0
